Added the main functionality
Created to successfully add and display the products.
Admin page, where he can manage the reports to the system or the questions from the users.
Created "Contact" page for the users where they have communication with the system.
For the products, made a "Shopping Cart" page which leads to the "Check-out" page with the sum of the products. Made simple form if the user wants to pay with card